+++ packages/jacaranda/src/routers/module.js
@@ -51,8 +51,6 @@ async function moduleRouter(app, baseRoute, options) {
             return;
         }
 
-        
-
         if (!Array.isArray(controllers)) {
             controllers = [controllers];
         }
@@ -86,13 +84,16 @@ async function moduleRouter(app, baseRoute, options) {
                 if (typeof action !== 'function' || !action.__metaHttpMethod) continue; // only marked httpMethod should be mounted
 
                 const method = action.__metaHttpMethod;
-                const subRoute = text.ensureStartsWith(urlUtil.join(
-                    subBaseRoute,
-                    text.dropIfStartsWith(
-                        action.__metaRoute || (kebabify ? naming.kebabCase(actionName) : actionName),
-                        '/'
-                    )
-                ), '/');
+                const subRoute = text.ensureStartsWith(
+                    urlUtil.join(
+                        subBaseRoute,
+                        text.dropIfStartsWith(
+                            action.__metaRoute || (kebabify ? naming.kebabCase(actionName) : actionName),
+                            '/'
+                        )
+                    ),
+                    '/'
+                );
 
                 let bindAction;