+++ packages/jacaranda/test/fastServer.js
@@ -1,0 +1,15 @@
+import { fastServe, http } from '../src';
+
+const modules = {
+    '/api': {
+        hello: http.Get(async (ctx) => {
+            ctx.body = {
+                hello: 'Jacaranda',
+            };
+        }),
+    },
+};
+
+const server = fastServe(modules);
+
+export default server;